refactor(ingest): isolate docs document normalization
