+++ klai-knowledge-ingest/knowledge_ingest/rebuild_tasks.py
@@ -281,8 +281,9 @@ async def _rebuild_artifact(
 
             # Step 1: Re-chunk with parent-child chunking (SPEC-RAG-PARENT-CHILD-001).
             from knowledge_ingest import chunker
+            from knowledge_ingest.document_normalizer import normalize_document_for_chunking
 
-            document_text = chunker.normalize_document_for_chunking(document_text)
+            document_text = normalize_document_for_chunking(document_text)
             child_chunks, parent_chunks_obj = chunker.chunk_markdown_with_parents(document_text)
             child_texts = [c.text for c in child_chunks]
             # Each child knows its parent's index in parent_chunks_obj —
